+++ MonsterVariants+/SubClasses/MultiplyGold.cs
@@ -7,7 +7,7 @@ namespace MonsterVariantsPlus.SubClasses
 {
     public class MultiplyGold
     {
-        private static int RewardMultiplier = 1;
+        private static float RewardMultiplier = 1f;
         public static uint MultiplyMoney(uint monsterGold, VariantHandler enemyVariant)
         {
             if (RunArtifactManager.instance.IsArtifactEnabled(Artifact.Variance))
